requirements.txt
Django>=5.1.3
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-nested-routers>=0.93.4